# server/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Travel Website API",
    "description": "API documentation for the travel website",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/cart/add": {
      "post": {
        "summary": "Add a trip to the cart",
        "description": "This endpoint adds a trip to the user's cart. If the trip is already in the cart, it increments the quantity.",
        "operationId": "addToCart",
        "tags": [
          "Cart"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tripId": {
                    "type": "string",
                    "description": "The unique ID of the trip to add to the cart"
                  },
                  "quantity": {
                    "type": "integer",
                    "default": 1,
                    "description": "The quantity of the trip to add to the cart"
                  }
                },
                "required": [
                  "tripId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trip successfully added to the cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Trip added to cart"
                    },
                    "cart": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string",
                          "description": "The user's ID"
                        },
                        "trips": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tripId": {
                                "type": "string",
                                "description": "The unique ID of the trip"
                              },
                              "quantity": {
                                "type": "integer",
                                "description": "The quantity of the trip in the cart"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          }
        }
      }
    },
    "/cart": {
      "get": {
        "summary": "Get the user's cart details",
        "description": "This endpoint retrieves the details of the user's cart.",
        "operationId": "getCart",
        "tags": [
          "Cart"
        ],
        "responses": {
          "200": {
            "description": "Cart details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "description": "The user's ID"
                    },
                    "trips": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tripId": {
                            "type": "string",
                            "description": "The unique ID of the trip"
                          },
                          "quantity": {
                            "type": "integer",
                            "description": "The quantity of the trip in the cart"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User needs to be authenticated"
          }
        }
      },
      "delete": {
        "summary": "Clear the user's entire cart",
        "description": "This endpoint clears all trips from the user's cart.",
        "operationId": "clearCart",
        "tags": [
          "Cart"
        ],
        "responses": {
          "200": {
            "description": "Cart cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Cart cleared successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User needs to be authenticated"
          }
        }
      }
    },
    "/cart/{tripId}": {
      "delete": {
        "summary": "Remove a trip from the cart",
        "description": "This endpoint removes a specific trip from the user's cart.",
        "operationId": "removeFromCart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "description": "The ID of the trip to remove from the cart",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trip removed from cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Trip removed from cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Trip not found in the cart"
          },
          "401": {
            "description": "Unauthorized - User needs to be authenticated"
          }
        }
      }
    },
...
  },
...
}